DualSpeechLM/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.3
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
datasets==3.0.0
deepspeed==0.15.1
diffusers==0.20.2
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
fairseq==0.12.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
flash-attn==2.6.3
Flask==3.0.3
fonttools==4.54.0
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gradio==3.50.0
gradio_client==0.6.1
grpcio==1.66.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lit==18.1.8
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.8.2
networkx==3.1
ninja==1.11.1.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.10.7
packaging==24.1
pandas==2.0.3
peft==0.4.0
pillow==10.4.0
pkgutil_resolve_name==1.3.10
portalocker==2.10.1
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.1.4
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.20.3
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchdata==0.8.0
torchvision==0.19.1+cu118
tqdm==4.66.5
transformers==4.45.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
webdataset==0.2.100
websockets==11.0.3
Werkzeug==3.0.4
xformers==0.0.28.post1
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
